requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-dotenv
python-docx
pymupdf
openai>=1.6.0
